requirements.txt
Flask
gevent-websocket
flask_sockets
openai-whisper